Separated FullpageController.svelte from Fullpage.svelte
